==> code/2_analyze_degs.Rmd <==

# libraries

```{r}
library(tidyverse)
source("functions.r")
library(openxlsx)
library(clusterProfiler)
library(org.Hs.eg.db)

rename <- dplyr::rename
select <- dplyr::select
```

# udf

```{r}
plot_venn = function(venn_obj, title = NULL) {
  pl = 
    RVenn::ggvenn(venn_obj,  fill = c("red2", "blue2", "green2")) +
  expand_limits(x = c(-2,6),
                y = c(-2.1,3)) +
  theme_void() +
  theme(legend.position = "none")
  
  if (!is.null(title)) {
    pl = pl + labs(title = title)
  }
  
  return(pl)
}
```

# import

```{r}
genes_list = readallsheets("../data/results/degs_cl.xlsx")[1:3]
```

Use gprofiler to convert gene names

```{r}
df_degs_comps_exp = 
  genes_list %>% 
  lapply(function(x) x %>% 
           filter(result != "NS") %>% 
           select(gene, comparison)) %>% 
  bind_rows() %>% 
  mutate(comparison = case_when(str_detect(comparison, "cl1") ~ "S1",
                                str_detect(comparison, "cl2") ~ "S2",
                                str_detect(comparison, "cl3") ~ "S3")) %>% 
  group_by(gene) %>% 
  summarise(comparison = paste0(unique(comparison), collapse = ", ")) %>% 
  arrange(comparison, gene)

df_gconv_query =
  gprofiler2::gconvert(query = df_degs_comps_exp$gene,
                       organism = "hsapiens")

df_degs_comps_exp_xl = 
  df_degs_comps_exp %>% 
  left_join(df_gconv_query %>% 
              rename(gene = input) %>% 
              select(gene, name, description),
            by = "gene") %>% 
  mutate(description = gsub("\\s*\\[.*?\\]", "", description)) %>% 
  relocate(comparison, .before = everything())

write.xlsx(df_degs_comps_exp_xl, "../data/results/degs/gprofiler_conv_degs_tab.xlsx")
```


# convert gene names

## bitr

Uso bitr dopo averlo confrontato con BioMart, perché il primo funziona meglio.

```{r}
bitr_conv_list = 
  genes_list %>% 
  lapply(function(x) bitr(x$gene,
                          fromType = "ENSEMBL",
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)) %>% 
  lapply(function(x) x %>% rename(gene = ENSEMBL))

genes_conv_list = vector("list", length = length(genes_list))
for (i in 1:length(genes_list)) {
  genes_conv_list[[i]] = 
    left_join(
      genes_list[[i]],
      bitr_conv_list[[i]],
      by = "gene"
    ) %>% 
    relocate(ENTREZID, .after = gene)
}
names(genes_conv_list) = names(genes_list)
```

# create input lists

```{r}
input_degs_list =
  genes_conv_list %>%
  lapply(function(x) x %>%
           na.omit() %>% 
           filter(my_padj < 0.1) %>%
           .$ENTREZID)
names(input_degs_list) = names(genes_list)

input_degs_ensg_list =
  genes_conv_list %>%
  lapply(function(x) x %>%
           na.omit() %>% 
           filter(my_padj < 0.1) %>%
           .$gene)
names(input_degs_ensg_list) = names(genes_list)
# input_degs_list %>% lapply(is.na) %>% lapply(table)
```

# ORA

## GO

```{r}
ora_go_list = input_degs_list # ne copia la struttura
for (i in 1:length(input_degs_list)) {
  ora_go_list[[i]] =
    enrichGO(input_degs_list[[i]], 
             OrgDb = org.Hs.eg.db,
             ont = "BP",
             maxGSSize = 50000,
             universe = genes_conv_list[[i]]$ENTREZID,
             readable = T)
}

ora_go_list %>% lapply(as.data.frame)
```

### venn

```{r}
ora_go_list %>%
  lapply(as.data.frame) %>% 
  lapply(function(x) x %>% pull(ID)) %>% 
  `names<-`(c("S1", "S2", "S3")) %>% 
  RVenn::Venn() %>% 
  plot_venn()# +
  #labs(title = "Gene Ontology")

ggsave("../data/results/ora/go_venn.png", 
       bg = "white",
       width = 15,
       height = 15,
       units = "cm",
       dpi = 300)
```


```{r}
ora_go_list %>%
  lapply(as.data.frame) %>% 
  lapply(function(x) x %>% pull(ID)) %>% 
  RVenn::Venn() %>% 
  RVenn::overlap_pairs(slice = c(1,3))

ora_go_list %>%
  lapply(as.data.frame) %>% 
  bind_rows(.id = "cluster") %>% 
  select(cluster, ID, Description, Count) %>% 
  `rownames<-`(NULL) %>% 
  filter(ID %in% (
    ora_go_list %>%
      lapply(as.data.frame) %>% 
      lapply(function(x) x %>% pull(ID)) %>% 
      RVenn::Venn() %>% 
      RVenn::overlap_pairs(slice = c(1,3)) %>%
      as.data.frame() %>% 
      pull(1) %>%
      as.character()
  ) 
  ) %>% 
  arrange(ID)
```


### export

```{r}
openxlsx::write.xlsx(ora_go_list %>% lapply(as.data.frame), "../data/results/ora/go.xlsx")
```


## KEGG

```{r}
ora_kegg_list = input_degs_list # ne copia la struttura
for (i in 1:length(input_degs_list)) {
  ora_kegg_list[[i]] =
    enrichKEGG(input_degs_list[[i]], 
               organism = "hsa",
               maxGSSize = 50000,
               universe = genes_conv_list[[i]]$ENTREZID)
}

ora_kegg_list %>% lapply(as.data.frame)
```

### venn

```{r}
ora_kegg_list %>%
  lapply(as.data.frame) %>% 
  lapply(function(x) x %>% pull(ID)) %>% 
  `names<-`(c("S1", "S2", "S3")) %>% 
  RVenn::Venn() %>% 
  plot_venn() #+
  # labs(title = "KEGG")

ggsave("../data/results/ora/kegg_venn.png", 
       bg = "white",
       width = 15,
       height = 15,
       units = "cm",
       dpi = 300)
```


### export

```{r}
openxlsx::write.xlsx(ora_kegg_list %>% lapply(as.data.frame), "../data/results/ora/kegg.xlsx")
```

## WikiPathways

```{r}
ora_wp_list = input_degs_list # ne copia la struttura
for (i in 1:length(input_degs_list)) {
  ora_wp_list[[i]] =
    enrichWP(input_degs_list[[i]], 
             organism = "Homo sapiens",
             maxGSSize = 50000,
             universe = genes_conv_list[[i]]$ENTREZID)
}

ora_wp_list %>% lapply(as.data.frame)
```

### venn

```{r}
ora_wp_list %>%
  lapply(as.data.frame) %>% 
  lapply(function(x) x %>% pull(ID)) %>% 
    `names<-`(c("S1", "S2", "S3")) %>% 
  RVenn::Venn() %>% 
  plot_venn() #+
  # labs(title = "WikiPathways")

ggsave("../data/results/ora/wp_venn.png", 
       bg = "white",
       width = 15,
       height = 15,
       units = "cm",
       dpi = 300)
```

### export

```{r}
openxlsx::write.xlsx(ora_wp_list %>% lapply(as.data.frame), "../data/results/ora/wp.xlsx")
```
