==> code/4_1_clusters_degs.Rmd <==

# libraries
```{r include=FALSE}
library(DESeq2)
library(tidyverse)
library(tidymodels)

library(broom)

source("functions.r")

rename <- dplyr::rename
select <- dplyr::select
```

# import
```{r}
df_clusters = read.csv("../data/repos/PPMI-Subtype-Analysis-master/subtype/input/clustering_by_lstm.csv")

load("../data/input/dds_objs_de.RData")
```


# coldata

```{r}
df_rna_metadata = 
  coldata_list_wov02 %>% 
  bind_rows() %>% 
  filter(cluster != "HC",
         time == "BL") %>% 
  droplevels()

df_rna_metadata %>% 
  pull(cluster) %>%
  table
```

# TRAIN TEST SPLIT

```{r}

set.seed(42) # for reproducibility

split <- 
  initial_split(df_rna_metadata %>% 
                  select(patno, cluster) %>% 
                  distinct() %>% 
                  droplevels(),
                strata = cluster,
                prop = 3/4)

# Extract the training and testing data
train_data <- training(split)
test_data <- testing(split)

train_data %>% 
  group_by(cluster) %>% 
  tally()
test_data %>% 
  group_by(cluster) %>% 
  tally()
```

# vst

on train and test

## coldata

```{r}
df_rna_metadata_train = 
  df_rna_metadata %>% 
  filter(patno %in% train_data$patno)

df_rna_metadata_test = 
  df_rna_metadata %>% 
  filter(patno %in% test_data$patno)
```


## countdata

```{r}
countdata_train =
  countdata_list_wov02 %>%
  lapply(
    function(x) x %>%
      rownames_to_column(var = "gene") %>%
      pivot_longer(cols = contains("_"))
  ) %>%
  bind_rows() %>% 
  filter(name %in% df_rna_metadata_train$names) %>% 
  pivot_wider(names_from = name,
              values_from = value) %>% 
  column_to_rownames("gene")

countdata_test =
  countdata_list_wov02 %>%
  lapply(
    function(x) x %>%
      rownames_to_column(var = "gene") %>%
      pivot_longer(cols = contains("_"))
  ) %>%
  bind_rows() %>% 
  filter(name %in% df_rna_metadata_test$names) %>% 
  pivot_wider(names_from = name,
              values_from = value) %>% 
  column_to_rownames("gene")
```


## dds

Questo DDS viene costruito apposta per calcolare VST.

```{r}
dds_train = 
  DESeqDataSetFromMatrix(
    countData = countdata_train,
    colData = df_rna_metadata_train,
    design = ~ 1 
  )

dds_train %>% design()

dds_test = 
  DESeqDataSetFromMatrix(
    countData = countdata_test,
    colData = df_rna_metadata_test,
    design = ~ 1
  )

dds_test %>% design()
```

## trans

```{r}
# run DESeq function on train data
dds_train_deseq = DESeq(dds_train, parallel = T)

# apply VST
dds_train_trans = varianceStabilizingTransformation(dds_train_deseq, blind = F)

# run DESeq function on test data
dds_test_deseq = DESeq(dds_test, parallel = T)

# copy VST properties
dispersionFunction(dds_test_deseq) = dispersionFunction(dds_train_deseq)

# apply VST
dds_test_trans = varianceStabilizingTransformation(dds_test_deseq, blind = F)
```

# feature selection

```{r}
df_anova = 
  dds_train_trans %>% 
  assay() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  pivot_longer(cols = contains("_")) %>% 
  left_join(df_clusters %>% 
              mutate(name = paste0(as.character(PATNO), "_BL"),
                     group = as.factor(CLUSTER_IDX),
                     .keep = "unused"),
            by = "name")
```

## export

```{r}
saveRDS(df_anova, "../data/input/classifier/train_rna_data.rds")

dds_test_trans %>% 
  assay() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  pivot_longer(cols = contains("_")) %>% 
  left_join(df_clusters %>% 
              mutate(name = paste0(as.character(PATNO), "_BL"),
                     group = as.factor(CLUSTER_IDX),
                     .keep = "unused"),
            by = "name") %>% 
  saveRDS("../data/input/classifier/test_rna_data.rds")
```

ANOVA

```{r}
df_anova_list = 
  df_anova %>% 
  split(f = .$gene)

library(parallel)
# create a cluster object with desired number of cores
cl <- makeCluster(detectCores()-1)

clusterCall(cl = cl,
            function() {
              #library(clusterProfiler)
              library(tidyverse)
              set.seed(217)
            } )

# apply gseGO function to each chunk in parallel using parLapply
df_anova_list_res <-
  df_anova_list %>%
  parLapply(cl = cl,
            X = .,
            function(x) {
              aov(formula = value ~ group,
                  data = x) %>% 
                broom::tidy() %>% 
                na.omit()
            })

# stop the cluster
stopCluster(cl)

# df_anova_list_res
```

extract results

```{r}
df_anova_res = 
  df_anova_list_res %>% 
  bind_rows(.id = "gene") %>% 
  mutate(p_adj = p.adjust(p = p.value, method = "bonferroni"),
         p_adj_fdr =  p.adjust(p = p.value, method = "fdr")) %>% 
  mutate(p_sig = p.value < 0.05,
         p_adj_sig = p_adj < 0.05,
         p_fdr_sig = p_adj_fdr < 0.05)

df_anova_res$p_adj_sig %>% table()
df_anova_res$p_fdr_sig %>% table()

df_anova_res %>% 
  arrange(p_adj_fdr)

df_selected_genes_train = 
  df_anova %>% 
  filter(gene %in% (df_anova_res %>% 
                      filter(p_fdr_sig) %>% 
                      pull(gene)))
  
```

## viz

```{r}
df_selected_genes_train %>% 
  ggplot() +
  aes(x = value,
      fill = group) +
  geom_density(alpha = .3) +
  theme_bw() +
  facet_wrap(~gene, scales = "free")
```

```{r}
df_selected_genes_test = 
  dds_test_trans %>% 
  assay() %>% 
  as.data.frame() %>% 
  # head() %>%
  rownames_to_column(var = "gene") %>% 
  # filter(!str_detect(gene, "PAR_Y")) %>% 
  # mutate(gene = sub("\\..*", "", gene)) %>% 
  pivot_longer(cols = contains("_")) %>% 
  left_join(df_clusters %>% 
              mutate(name = paste0(as.character(PATNO), "_BL"),
                     group = as.factor(CLUSTER_IDX),
                     .keep = "unused"),
            by = "name") %>% 
  filter(gene %in% (df_anova_res %>% 
                      filter(p_fdr_sig) %>% 
                      pull(gene)))
```

```{r}
df_selected_genes_test %>% 
  ggplot() +
  aes(x = value,
      fill = group) +
  geom_density(alpha = .3) +
  theme_bw() +
  facet_wrap(~gene, scales = "free")
```

# export

```{r}
openxlsx::write.xlsx(df_anova_res, file = "../data/input/classifier/anova_res.xlsx")

save(df_selected_genes_train,
     df_selected_genes_test,
     file = "../data/input/classifier/selected_genes_data_train_e_test.rdata")
```
